Fix: Type coercion issues after MySQL + Redis migration
Fixed TypeError in CircuitBreaker and missing cast in Order model caused
by migrating from SQLite to MySQL + Redis cache.

Root cause:
- SQLite returns integers as int type
- MySQL + Redis return integers as string type
- PHP 8.2+ strict typing catches these mismatches

Fixes:
1. CircuitBreaker (app/Services/Linnworks/Core/CircuitBreaker.php):
   - Added (int) casts to Cache::get() for failure/success counts
   - Ensures getFailureCount() and getSuccessCount() return int not string
   - Fixed incrementFailureCount() and incrementSuccessCount()

2. Order model (app/Models/Order.php):
   - Added missing 'is_paid' => 'boolean' cast
   - Ensures is_paid field returns bool not string/int

Result: No more "Return value must be of type int, string returned" errors

Technical note:
Redis stores all values as strings. When retrieving numeric values from
Redis cache, explicit casting to int/float is required for strict typing.
Laravel model casts handle this automatically for Eloquent, but manual
casts needed for direct Cache::get() calls.

+++ app/Services/Linnworks/Core/CircuitBreaker.php
@@ -188,7 +188,7 @@ class CircuitBreaker
     private function incrementFailureCount(): int
     {
         $key = $this->getFailureCountKey();
-        $count = Cache::get($key, 0) + 1;
+        $count = (int) Cache::get($key, 0) + 1;
         Cache::put($key, $count, now()->addHours(24));
         Cache::put($this->getLastFailureKey(), now(), now()->addHours(24));
 
@@ -201,7 +201,7 @@ class CircuitBreaker
     private function incrementSuccessCount(): int
     {
         $key = $this->getSuccessCountKey();
-        $count = Cache::get($key, 0) + 1;
+        $count = (int) Cache::get($key, 0) + 1;
         Cache::put($key, $count, now()->addHours(24));
 
         return $count;
@@ -228,7 +228,7 @@ class CircuitBreaker
      */
     private function getFailureCount(): int
     {
-        return Cache::get($this->getFailureCountKey(), 0);
+        return (int) Cache::get($this->getFailureCountKey(), 0);
     }
 
     /**
@@ -236,7 +236,7 @@ class CircuitBreaker
      */
     private function getSuccessCount(): int
     {
-        return Cache::get($this->getSuccessCountKey(), 0);
+        return (int) Cache::get($this->getSuccessCountKey(), 0);
     }
 
     private function getStateKey(): string
